==> analysis/02-analysis.Rmd <==
---
title: "Analysis 2"
author: "James Ashmore"
date: "2019-10-23"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Setup

Set knitr chunk options:

```{r knitr}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/02-analysis.Rmd/",
  cache.lazy = FALSE,
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

## Packages

Load required analysis packages:

```{r, cache = FALSE}
pacman::p_load(batchelor, gam, here, readr, scater, scran, slingshot, RColorBrewer, pheatmap, BiocParallel)
```

## Parallelization

```{r}
bpparam <- MulticoreParam(4)
```

## Atlas

Load the MouseGastrulationData package:

```{r, cache = FALSE}
pacman::p_load(MouseGastrulationData)
```

Modify colour palette to include unassigned and unknown cells:

```{r, cache = FALSE}
EmbryoCelltypeColours <- c(EmbryoCelltypeColours, unassigned = "#999999", unknown = "#999999")
```

Select samples from embryonic stage E8.25:

```{r, cache = FALSE}
sel <- subset(AtlasSampleMetadata, stage %in% c("E7.0", "E7.25", "E7.5", "E7.75"))
```

Obtain processed counts from the dataset:

```{r}
ref <- EmbryoAtlasData(type = "processed", samples = sel$sample)
```

Remove cells with no defined cell type:

```{r, cache = FALSE}
nan <- is.na(ref$celltype)
ref <- ref[, !nan]
```

Compute log-transformed normalized expression values:

```{r}
ref <- logNormCounts(ref)
```

Change dimension names to those required by the scater package:

```{r, cache = FALSE}
reducedDimNames(ref) <- c("PCA", "UMAP")
```

Plot UMAP showing all the selected cells: 

```{r, cache = FALSE}
plt <- plotUMAP(ref)

plt$data$colour_by <- ref$celltype

plt$data$facet_by <- ref$stage

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(colour = colour_by)) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_colour_manual(name = "Cell type", values = EmbryoCelltypeColours) +
  theme_void() +
  theme(aspect.ratio = 1)
```

## EHF

Load the EHF dataset:

```{r}
sce <- read_rds(here("data/01-analysis.rds"))
```

## Neighbors

Clear spike-in information across batches:

```{r, cache = FALSE}
ref <- clearSpikes(ref)

sce <- clearSpikes(sce)
```

Filter batches to contain same genes:

```{r, cache = FALSE}
rownames(ref) <- rowData(ref)$ENSEMBL

rownames(sce) <- rowData(sce)$gene_id

ids <- intersect(rownames(ref), rownames(sce))

ref <- ref[ids, ]

sce <- sce[ids, ]
```

Select genes with positive average biological components:

```{r}
mod <- model.matrix(~ stage + sample, data = colData(ref))

var.ref <- modelGeneVar(ref, design = mod, BPPARAM = bpparam)
  
var.sce <- modelGeneVar(sce, block = sce$phase, BPPARAM = bpparam)

dec <- combineVar(var.ref, var.sce)

hvg <- rownames(dec)[dec$bio > 0]
```

Rename cells from query and reference data:

```{r, cache = FALSE}
colnames(sce) <- paste0("wilson-", seq_len(ncol(sce)))

colnames(ref) <- paste0("marioni-", seq_len(ncol(ref)))
```

Correct for batch effects using a fast version of the MNN method:

```{r}
mnn <- fastMNN(ref, sce, pc.input = TRUE, subset.row = hvg, BPPARAM = bpparam)

reducedDimNames(mnn) <- "MNN"
```

Incorporate raw and log counts from original data:

```{r, cache = FALSE}
assay(mnn, "counts") <- cbind(assay(ref[hvg, ], "counts"), assay(sce[hvg, ], "counts"))

assay(mnn, "logcounts") <- cbind(assay(ref[hvg, ], "logcounts"), assay(sce[hvg, ], "logcounts"))
```

Annotate merged dataset with metadata:

```{r, cache = FALSE}
#Batch
mnn$batch <- ifelse(mnn$batch == 1, "Marioni", "Wilson")

# Celltype
mnn$celltype <- c(ref$celltype, sce$cellassign_celltype)

# Pool
mnn$stage <- "NA"
mnn$stage[mnn$batch == "Marioni"] <- ref$stage

# Sample
mnn$sample <- "NA"
mnn$sample[mnn$batch == "Marioni"] <- ref$sample

# Phase
mnn$phase <- "NA"
mnn$phase[mnn$batch == "Wilson"] <- as.character(sce$phase)

# Phase
mnn$germ <- "NA"
mnn$germ[mnn$batch == "Wilson"] <- as.character(sce$germ_layer)
```

Choose optimal K from reference cells:

```{r, eval = FALSE}
dim <- reducedDim(mnn, "MNN")[mnn$batch == "Marioni", ]

colnames(dim) <- paste0("PC", seq_len(ncol(dim)))

dim <- as.data.frame(dim)

fct <- paste0(mnn$stage, "-", mnn$celltype)[mnn$batch == "Marioni"]

mod <- caret::trainControl(
  method = "LGOCV", 
  p = 0.7,
  number = 1,
  savePredictions = TRUE
)

grd <- expand.grid(k = seq(1, 50, by = 2))

set.seed(1701)

fit <- caret::train(x = dim, y = fct, method = "knn", tuneGrid = grd, trControl = mod)

num <- fit$bestTune$k
```

Identify kNN from reference cells:

```{r, cache = FALSE}
set.seed(1701)

num <- 2

sbj <- reducedDim(mnn, "MNN")[mnn$batch == "Marioni", ]

qry <- reducedDim(mnn, "MNN")[mnn$batch == "Wilson", ]

knn <- BiocNeighbors::queryKNN(sbj, qry, k = num)
```

Annotate cell type using kNN cells:

```{r, cache = FALSE}
nan <- rep(NA, ncol(sce))

# Neighbor

mnn$neighbor <- c(ref$celltype, nan)

lab <- apply(knn$index, 1, function(x) names(which.max(table(ref$celltype[x]))))

mnn$neighbor[mnn$batch == "Wilson"] <- lab

# Stage

mnn$stage <- c(ref$stage, nan)

lab <- apply(knn$index, 1, function(x) names(which.max(table(ref$stage[x]))))

mnn$stage[mnn$batch == "Wilson"] <- lab
```

Filter merged dataset to k closest cells:

```{r, cache = FALSE}
id1 <- which(mnn$batch == "Wilson")

len <- seq_len(num)

id2 <- as.vector(knn$index[, len])

idx <- c(id1, id2)

idx <- unique(sort(idx))

mnn <- mnn[, idx]
```

Perform UMAP on the PCA data from the combined dataset:

```{r, cache = FALSE}
set.seed(1701)

mnn <- runUMAP(mnn, n_neighbors = 50, dimred = "MNN", BPPARAM = bpparam)
```

Plot UMAP showing batch annotation:

```{r, cache = FALSE}
plt <- plotUMAP(mnn)

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

col <- c("Marioni" = "#d9d9d9", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(colour = colour_by)) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_colour_manual(name = "Batch", values = col) +
  theme_void() +
  theme(aspect.ratio = 1)
```

Plot UMAP showing celltype annotation:

```{r, cache = FALSE}
plt <- plotUMAP(mnn)

plt$data$fill_by <- mnn$celltype

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

col <- setdiff(unique(mnn$celltype), unique(mnn$neighbor))

col <- RColorBrewer::brewer.pal(length(col), name = "Set1")

names(col) <- setdiff(unique(mnn$celltype), unique(mnn$neighbor))

col <- c(EmbryoCelltypeColours, col)

pch <- c("Marioni" = "transparent", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(fill = fill_by, colour = colour_by), shape = 21) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_fill_manual(name = "Cell type", values = col) + 
  scale_colour_manual(name = "Batch", values = pch, limits = "Wilson") + 
  theme_void() +
  theme(aspect.ratio = 1)
```

Plot UMAP showing germ layer annotation:

```{r, cache = FALSE}
plt <- plotUMAP(mnn)

plt$data$fill_by <- mnn$germ

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

col <- c(RColorBrewer::brewer.pal(3, name = "Set1"), "#d9d9d9")

names(col) <- c("ectoderm", "intermediate", "mesoderm", "NA")

pch <- c("Marioni" = "transparent", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(fill = fill_by, colour = colour_by), shape = 21) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_fill_manual(name = "Cell type", values = col) + 
  scale_colour_manual(name = "Batch", values = pch, limits = "Wilson") + 
  theme_void() +
  theme(aspect.ratio = 1)
```

Plot UMAP showing neighbour and batch annotation:

```{r, cache = FALSE}
plt <- plotUMAP(mnn)

plt$data$fill_by <- mnn$neighbor

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

col <- EmbryoCelltypeColours

pch <- c("Marioni" = "transparent", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(fill = fill_by, colour = colour_by), shape = 21) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_fill_manual(name = "Cell type", values = col) + 
  scale_colour_manual(name = "Batch", values = pch, limits = "Wilson") + 
  theme_void() +
  theme(aspect.ratio = 1)
```

Save merged dataset for further exploration:

```{r}
write_rds(mnn, here("data/02-analysis.rds"))
```
