==> analysis/04-analysis.Rmd <==
---
title: "Analysis 4"
author: "James Ashmore"
date: "2019-10-23"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Setup

Set knitr chunk options:

```{r knitr}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/04-analysis.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

## Packages

Load required analysis packages:

```{r}
pacman::p_load(batchelor, gam, here, readr, scater, scran, slingshot, RColorBrewer, pheatmap, igraph, EnsDb.Mmusculus.v79)
```

## Setup

Load the MouseGastrulationData package:

```{r}
pacman::p_load(MouseGastrulationData)
```

Modify colour palette to include unassigned and unknown cells:

```{r}
EmbryoCelltypeColours <- c(EmbryoCelltypeColours, unassigned = "#999999", unknown = "#999999")
```

## Import

Load the merged dataset:

```{r}
mnn <- read_rds(here("data/02-analysis.rds"))
```

Annotate gene identifiers:

```{r}
rowData(mnn)$ENSEMBL <- rownames(mnn)

rowData(mnn)$SYMBOL <- mapIds(EnsDb.Mmusculus.v79, keys = rownames(mnn), column = "SYMBOL", keytype = "GENEID")

rownames(mnn) <- uniquifyFeatureNames(rowData(mnn)$ENSEMBL, rowData(mnn)$SYMBOL)
```

Plot UMAP showing neighbour and batch annotation:

```{r}
plt <- plotUMAP(mnn)

plt$data$fill_by <- mnn$neighbor

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

col <- EmbryoCelltypeColours

pch <- c("Marioni" = "transparent", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(fill = fill_by, colour = colour_by), shape = 21) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_fill_manual(name = "Cell type", values = col) + 
  scale_colour_manual(name = "Batch", values = pch, limits = "Wilson") + 
  theme_void() +
  theme(aspect.ratio = 1)
```

Plot UMAP showing germ layer and batch annotation:

```{r}
plt <- plotUMAP(mnn)

plt$data$fill_by <- mnn$germ

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

col <- c(RColorBrewer::brewer.pal(3, name = "Set1"), "#d9d9d9")

names(col) <- c("ectoderm", "intermediate", "mesoderm", "NA")

pch <- c("Marioni" = "transparent", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) +
  geom_point(aes(fill = fill_by, colour = colour_by), shape = 21) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_fill_manual(name = "Cell type", values = col) + 
  scale_colour_manual(name = "Batch", values = pch, limits = "Wilson") + 
  theme_void() +
  theme(aspect.ratio = 1)
```

```{r}
snn <- buildSNNGraph(mnn, assay.type = "reconstructed", use.dimred = "MNN")

cls <- igraph::cluster_walktrap(snn)$membership

mnn$cluster_full <- factor(cls)
```

```{r, fig.width = 10, fig.height = 6}
plt <- plotUMAP(mnn)

plt$data$fill_by <- mnn$cluster_full

plt$data$colour_by <- mnn$batch

plt$data$facet_by <- mnn$stage

pch <- c("Marioni" = "transparent", "Wilson" = "#000000")

ggplot(plt$data, aes(X, Y)) + 
  geom_point(data = plt$data[, 1:4], mapping = aes(X, Y), colour = "grey") + 
  geom_point(aes(fill = fill_by, colour = colour_by), shape = 21) + 
  facet_wrap(~ facet_by, scales = "fixed") + 
  scale_fill_brewer(name = "Cluster", palette = "Set1") + 
  scale_colour_manual(name = "Batch", values = pch, limits = "Wilson") + 
  theme_void() +
  theme(aspect.ratio = 1)
```

```{r}
saveRDS(mnn, file = here("data/04-analysis.rds"))
```

## Sub-clustering

```{r}
sclust <- function(obj) {
  
  ref <- mnn[, mnn$batch == "Marioni"]
  
  sce <- mnn[, mnn$batch == "Wilson"]
  
  dec <- combineVar(modelGeneVar(ref), modelGeneVar(sce, block = sce$phase))
  
  hvg <- rownames(dec)[dec$bio > 0]
  
  snn <- buildSNNGraph(obj, assay.type = "reconstructed", use.dimred = "MNN", subset.row = hvg)
  
  cls <- igraph::cluster_walktrap(snn)$membership
  
  obj$cluster_walktrap <- factor(cls)
  
  return(obj)
  
}
```

```{r}
tbl <- table(mnn$cluster_full)

lvl <- names(tbl)

lvl <- lvl[tbl > 1]
```

```{r}
mnn <- lapply(lvl, function(x) {
  
  obj <- mnn[, mnn$cluster_full == x]
  
  obj <- sclust(obj)
  
  obj

})

names(mnn) <- lvl
```

```{r}
sig <- lapply(mnn, function(x) {
  
  m <- findMarkers(x, x$cluster_walktrap, block = x$batch, assay.type = "logcounts")
  
  m <- lapply(m, subset, FDR < 0.1)
  
  m <- lapply(m, head, n = 25)
  
  m <- lapply(m, rownames)
  
  m <- unique(sort(unlist(m)))

})

sig <- Filter(length, sig)
```

### Dimension {.tabset}

Create dimensionality reduction plots coloured by the top 5 marker genes:

```{r, fig.height = 4, fig.width = 6, results = "asis"}
for (n in names(sig)) {
  
  obj <- mnn[[n]]
  
  plt <- plotUMAP(obj, colour_by = "cluster_walktrap", shape_by = "batch")
  
  cat("#### Cluster", n, "\n")

  print(plt)

  cat("\n\n")

}
```

### Heatmap {.tabset}

Create a heatmap of expression values for the top 5 marker genes:

```{r, fig.height = 8, fig.width = 10, results = "asis"}
for (n in names(sig)) {

  obj <- mnn[[n]]
  
  ids <- sig[[n]]

  cat("#### Cluster", n, "\n")

  plotHeatmap(obj, ids, exprs_values = "reconstructed", center = TRUE, symmetric = TRUE, colour_columns_by = "batch", order_columns_by = "cluster_walktrap", show_colnames = FALSE, cluster_rows = ifelse(length(ids) >= 2, TRUE, FALSE))

  cat("\n\n")

}
```
